feat: Improve file synchronization and logging
This commit introduces several enhancements to file synchronization and logging mechanisms:

- Updates the spyrun.toml configuration file to include the latest version and to modify log synchronization parameters to ensure only files older than a day are moved.
- Adds event_path and event_kind to the log output for better debugging.
- Updates the sync_core_cfg, sync_system_cfg, sync_user_cfg, sync_system_cmd_global, sync_system_cmd_local, sync_user_cmd_global, sync_user_cmd_local, and sync_ng spy configurations to use debounce and limitkey to prevent running the same command multiple times.
- Introduces a new init.cmd to initialize the environment and synchronize local command files.
- Adds a new remove_task.cmd to clean up old task scheduler settings.
- Introduces new `gci_pdf.cmd`, `error_task.cmd`, and `oneshot.cmd` for various local command executions.
- Updates the archive.cmd to archive log files and remove old archive files.
- Updates move_log.cmd to use the test path.
- Adds a new sync_realtime.cmd to monitor file system changes and sync in real-time.
- Adds a new sync_sometime.cmd to monitor file system changes and sync sometime.
- Updates startup.cmd to start autohotkey, clnch and espanso at startup.
- Updates sync2onedrive.cmd to sync local user files to OneDrive.
- Adds logging to common.ps1 to show version and app info.
- Add minage option to robocopy for log sync.

+++ core/cfg/common.ps1
@@ -7,7 +7,7 @@
     - None
   .OUTPUTS
     - 0: SUCCESS / 1: ERROR
-  .Last Change: 2024/11/13 00:23:08.
+  .Last Change: 2024/12/05 00:16:04.
 #>
 $ErrorActionPreference = "Stop"
 $DebugPreference = "SilentlyContinue" # Continue SilentlyContinue Stop Inquire
@@ -531,6 +531,10 @@ function Exit-Task {
     log "[Exit-Task] Error ! result: [${result}]" "Red"
   }
 
+  if ([string]::IsNullOrEmpty($arg.xml)) {
+    return $result
+  }
+
   log "[Exit-Task] Remove task xml: $($arg.xml)"
   $result = Remove-ScheduledTask ([PSCustomObject]@{
       xml = $arg.xml
@@ -879,7 +883,7 @@ function Start-Init {
   $app.Add("resultPrefixFileLocal", $app.resultPrefixFileRemote.Replace($app.baseRemote, $app.baseLocal))
 
   if ($app.userType -eq "core") {
-    $app.Add("logDir", [System.IO.Path]::Combine($app.baseRemote, $app.userType, "log", $app.scope, $app.cmdName))
+    $app.Add("logDir", [System.IO.Path]::Combine($app.baseRemote, $app.userType, "log", $env:COMPUTERNAME, $app.scope, $app.cmdName))
   } else {
     $app.Add("logDir", [System.IO.Path]::Combine($app.baseLocal, $app.userType, "log", $app.scope, $app.cmdName))
   }
@@ -890,6 +894,7 @@ function Start-Init {
   Start-Transcript $app.logFile
 
   log "[Start-Init] version: $($app.version)"
+  log ([PSCustomObject]$app | ConvertTo-Json)
 
   # const value.
   $app.Add("cnst", @{
